Prevent the deletion of symlinks protected explicitly (bug #899721).
A symlink can now be protected if the symlink itself is mentioned in a config
file. However, a symlink which dereferences to a protected path will not be
protected.

This means that if you have a symlink like "/usr/lib64/lib -> /usr/lib", you
can put "/usr/lib64/lib" in your /etc/safe-rm.conf and then the following
will be skipped:

   rm /usr/lib64/lib

while the following will be deleted:

   ln -s /usr/lib64/lib /foo
   rm /foo